chore(lint): resolve golangci-lint findings
Fix all golangci-lint findings: canonical header names (X-Merchantid, Subscription-Userinfo), wrap errors with %w, check Body.Close and json encode/marshal errors, named interface params, nil-guard cfg in NewHandler (SA5011), drop minute truncation cleanup leftovers, handle ParseUint without nilerr, exitAfterDefer in main (explicit sentry flush before os.Exit), simplify migrations overflow guard (G115), rename db_pool_wait_total gauge to db_pool_wait (promlinter), and test-only cleanups: blank-line whitespace, dogsled, copyloopvar, dupword, unconvert, ineffassign.

+++ internal/database/migration_test.go
@@ -21,7 +21,7 @@ func TestMigrationVersionToInt(t *testing.T) {
 	}{
 		{name: "positive", version: 31, want: 31},
 		{name: "zero", version: 0, wantErr: true},
-		{name: "overflow", version: uint(^uint(0)>>1) + 1, wantErr: true},
+		{name: "overflow", version: ^uint(0)>>1 + 1, wantErr: true},
 	}
 
 	for _, tt := range tests {
